requirements.txt
requests
babel
pycoingecko
pymempool>=0.1.2
Pillow
matplotlib
numpy
pydantic
mplfinance
pandas
